Do not create new target groups before deleting old ones
The property create_before_delete is propagated to the aws_lb resource,
and causes failures when making changes on an LB which require the LB to
be recreated, e.g. changing the subnets in which the LB exists when
setting a fixed name using `name` (instead of `name_prefix`), or doing
the same when using `name_prefix` but providing externally managed
elastic IPs.